refactor: ComputerAnswer 클래스에서 컴퓨터의 정답 관련 필드와 메서드 만들기
setValue: 컴퓨터의 정답을 설정한다.
comparePlayerAnswer: 플레이어의 정답과 비교하여 볼과, 스트라이크의 개수를 반환한다.
resetValue: 컴퓨터의 정답을 재설정한다.
